Move pkg/io to prow/io
