+++ BUILD.bazel
@@ -121,7 +121,6 @@ filegroup(
         "//pkg/flagutil:all-srcs",
         "//pkg/genyaml:all-srcs",
         "//pkg/ghclient:all-srcs",
-        "//pkg/io:all-srcs",
         "//prow:all-srcs",
         "//releng:all-srcs",
         "//robots/commenter:all-srcs",
