+++ prow/cmd/gerrit/BUILD.bazel
@@ -10,12 +10,12 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//pkg/flagutil:go_default_library",
-        "//pkg/io:go_default_library",
         "//prow/config:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/gerrit/adapter:go_default_library",
         "//prow/gerrit/client:go_default_library",
         "//prow/interrupts:go_default_library",
+        "//prow/io:go_default_library",
         "//prow/logrusutil:go_default_library",
         "//prow/pjutil:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
@@ -56,8 +56,8 @@ go_test(
     srcs = ["main_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/io:go_default_library",
         "//prow/gerrit/client:go_default_library",
+        "//prow/io:go_default_library",
         "@com_google_cloud_go//storage:go_default_library",
         "@io_k8s_apimachinery//pkg/util/sets:go_default_library",
     ],
