+++ prow/flagutil/BUILD.bazel
@@ -16,7 +16,6 @@ go_library(
     importpath = "k8s.io/test-infra/prow/flagutil",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/io:go_default_library",
         "//prow/bugzilla:go_default_library",
         "//prow/client/clientset/versioned:go_default_library",
         "//prow/client/clientset/versioned/typed/prowjobs/v1:go_default_library",
@@ -24,6 +23,7 @@ go_library(
         "//prow/git:go_default_library",
         "//prow/git/v2:go_default_library",
         "//prow/github:go_default_library",
+        "//prow/io:go_default_library",
         "//prow/kube:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_k8s_client_go//kubernetes:go_default_library",
