+++ prow/io/BUILD.bazel
@@ -6,10 +6,10 @@ go_library(
         "opener.go",
         "option.go",
     ],
-    importpath = "k8s.io/test-infra/pkg/io",
+    importpath = "k8s.io/test-infra/prow/io",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/io/providers:go_default_library",
+        "//prow/io/providers:go_default_library",
         "@com_github_aws_aws_sdk_go//service/s3/s3manager:go_default_library",
         "@com_github_googlecloudplatform_testgrid//util/gcs:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
@@ -32,7 +32,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//pkg/io/providers:all-srcs",
+        "//prow/io/providers:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
