+++ prow/io/providers/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["providers.go"],
-    importpath = "k8s.io/test-infra/pkg/io/providers",
+    importpath = "k8s.io/test-infra/prow/io/providers",
     visibility = ["//visibility:public"],
     deps = [
         "@com_github_aws_aws_sdk_go//aws:go_default_library",
