+++ prow/pod-utils/gcs/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     importpath = "k8s.io/test-infra/prow/pod-utils/gcs",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/io:go_default_library",
         "//prow/apis/prowjobs/v1:go_default_library",
+        "//prow/io:go_default_library",
         "//prow/pod-utils/downwardapi:go_default_library",
         "@com_github_googlecloudplatform_testgrid//metadata:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
@@ -46,8 +46,8 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/io:go_default_library",
         "//prow/apis/prowjobs/v1:go_default_library",
+        "//prow/io:go_default_library",
         "//prow/pod-utils/downwardapi:go_default_library",
         "@io_k8s_apimachinery//pkg/api/equality:go_default_library",
         "@io_k8s_utils//pointer:go_default_library",
