+++ prow/pod-utils/gcs/upload.go
@@ -31,7 +31,7 @@ import (
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	utilpointer "k8s.io/utils/pointer"
 
-	pkgio "k8s.io/test-infra/pkg/io"
+	pkgio "k8s.io/test-infra/prow/io"
 )
 
 // UploadFunc knows how to upload into an object
