+++ prow/statusreconciler/BUILD.bazel
@@ -11,10 +11,10 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//maintenance/migratestatus/migrator:go_default_library",
-        "//pkg/io:go_default_library",
         "//prow/client/clientset/versioned/typed/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
         "//prow/github:go_default_library",
+        "//prow/io:go_default_library",
         "//prow/pjutil:go_default_library",
         "//prow/plugins:go_default_library",
         "//prow/plugins/trigger:go_default_library",
@@ -33,9 +33,9 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/io:go_default_library",
         "//prow/config:go_default_library",
         "//prow/github:go_default_library",
+        "//prow/io:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/util/diff:go_default_library",
