+++ prow/tide/BUILD.bazel
@@ -10,11 +10,11 @@ go_library(
     importpath = "k8s.io/test-infra/prow/tide",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/io:go_default_library",
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
         "//prow/git/v2:go_default_library",
         "//prow/github:go_default_library",
+        "//prow/io:go_default_library",
         "//prow/pjutil:go_default_library",
         "//prow/tide/blockers:go_default_library",
         "//prow/tide/history:go_default_library",
