+++ prow/tide/history/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "k8s.io/test-infra/prow/tide/history",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/io:go_default_library",
         "//prow/apis/prowjobs/v1:go_default_library",
+        "//prow/io:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
@@ -17,8 +17,8 @@ go_test(
     srcs = ["history_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/io:go_default_library",
         "//prow/apis/prowjobs/v1:go_default_library",
+        "//prow/io:go_default_library",
         "@com_google_cloud_go//storage:go_default_library",
         "@io_k8s_apimachinery//pkg/util/diff:go_default_library",
     ],
